AuthPortal/backend_authportal/pyproject.toml
[tool.poetry]
name = "backend-authportal"
version = "0.1.0"
description = ""
authors = ["Arman229 <armanashraf015@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standared"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
psycopg = {extras = ["binary"], version = "^3.1.18"}
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.7.1"}
passlib = "^1.7.4"
bcrypt = "^4.1.3"
types-python-jose = "^3.3.4.20240106"


[tool.poetry.scripts]
myproject = "backend_authportal.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
